gcloudfunctiongcscopy/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=2.9.0
functions-framework>=3.4.0
